+++ HorseFluteAnywhere/ModEntry.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Microsoft.Xna.Framework;
 using Pathoschild.Stardew.Common;
+using Pathoschild.Stardew.Common.Integrations.GenericModConfigMenu;
 using Pathoschild.Stardew.Common.Patching;
 using Pathoschild.Stardew.HorseFluteAnywhere.Framework;
 using Pathoschild.Stardew.HorseFluteAnywhere.Patches;
@@ -69,24 +70,12 @@ internal class ModEntry : Mod
     /// <inheritdoc cref="IGameLoopEvents.GameLaunched" />
     private void OnGameLaunched(object? sender, GameLaunchedEventArgs e)
     {
-        // add Generic Mod Config Menu integration
-        new GenericModConfigMenuIntegrationForHorseFluteAnywhere(
-            getConfig: () => this.Config,
-            reset: () =>
-            {
-                this.Config = new ModConfig();
-                this.Helper.WriteConfig(this.Config);
-                this.UpdateConfig();
-            },
-            saveAndApply: () =>
-            {
-                this.Helper.WriteConfig(this.Config);
-                this.UpdateConfig();
-            },
-            modRegistry: this.Helper.ModRegistry,
-            monitor: this.Monitor,
-            manifest: this.ModManifest
-        ).Register();
+        this.AddGenericModConfigMenu(
+            new GenericModConfigMenuIntegrationForHorseFluteAnywhere(),
+            get: () => this.Config,
+            set: config => this.Config = config,
+            onSaved: this.UpdateConfig
+        );
     }
 
     /// <inheritdoc cref="IInputEvents.ButtonsChanged" />