+++ NoclipMode/Framework/GenericModConfigMenuIntegrationForNoclipMode.cs
@@ -1,43 +1,19 @@
-using System;
 using Pathoschild.Stardew.Common.Integrations.GenericModConfigMenu;
 using StardewModdingAPI;
 
 namespace Pathoschild.Stardew.NoclipMode.Framework;
 
 /// <summary>Registers the mod configuration with Generic Mod Config Menu.</summary>
-internal class GenericModConfigMenuIntegrationForNoclipMode
+internal class GenericModConfigMenuIntegrationForNoclipMode : IGenericModConfigMenuIntegrationFor<ModConfig>
 {
-    /*********
-    ** Fields
-    *********/
-    /// <summary>The Generic Mod Config Menu integration.</summary>
-    private readonly GenericModConfigMenuIntegration<ModConfig> ConfigMenu;
-
-
     /*********
     ** Public methods
     *********/
-    /// <summary>Construct an instance.</summary>
-    /// <param name="modRegistry">An API for fetching metadata about loaded mods.</param>
-    /// <param name="monitor">Encapsulates monitoring and logging.</param>
-    /// <param name="manifest">The mod manifest.</param>
-    /// <param name="getConfig">Get the current config model.</param>
-    /// <param name="reset">Reset the config model to the default values.</param>
-    /// <param name="saveAndApply">Save and apply the current config model.</param>
-    public GenericModConfigMenuIntegrationForNoclipMode(IModRegistry modRegistry, IMonitor monitor, IManifest manifest, Func<ModConfig> getConfig, Action reset, Action saveAndApply)
+    /// <inheritdoc />
+    public void Register(GenericModConfigMenuIntegration<ModConfig> menu, IMonitor monitor)
     {
-        this.ConfigMenu = new GenericModConfigMenuIntegration<ModConfig>(modRegistry, monitor, manifest, getConfig, reset, saveAndApply);
-    }
-
-    /// <summary>Register the config menu if available.</summary>
-    public void Register()
-    {
-        var menu = this.ConfigMenu;
-        if (!menu.IsLoaded)
-            return;
-
-        menu.Register();
         menu
+            .Register()
             .AddSectionTitle(I18n.Config_Title_GeneralOptions)
             .AddCheckbox(
                 name: I18n.Config_EnabledMessage_Name,