+++ TractorMod/Framework/GenericModConfigMenuIntegrationForTractor.cs
@@ -7,14 +7,11 @@ using StardewModdingAPI;
 namespace Pathoschild.Stardew.TractorMod.Framework;
 
 /// <summary>Registers the mod configuration with Generic Mod Config Menu.</summary>
-internal class GenericModConfigMenuIntegrationForTractor
+internal class GenericModConfigMenuIntegrationForTractor : IGenericModConfigMenuIntegrationFor<ModConfig>
 {
     /*********
     ** Fields
     *********/
-    /// <summary>The Generic Mod Config Menu integration.</summary>
-    private readonly GenericModConfigMenuIntegration<ModConfig> ConfigMenu;
-
     /// <summary>An API for fetching metadata about loaded mods.</summary>
     private readonly IModRegistry ModRegistry;
 
@@ -24,28 +21,16 @@ internal class GenericModConfigMenuIntegrationForTractor
     *********/
     /// <summary>Construct an instance.</summary>
     /// <param name="modRegistry">An API for fetching metadata about loaded mods.</param>
-    /// <param name="monitor">Encapsulates monitoring and logging.</param>
-    /// <param name="manifest">The mod manifest.</param>
-    /// <param name="getConfig">Get the current config model.</param>
-    /// <param name="reset">Reset the config model to the default values.</param>
-    /// <param name="saveAndApply">Save and apply the current config model.</param>
-    public GenericModConfigMenuIntegrationForTractor(IModRegistry modRegistry, IMonitor monitor, IManifest manifest, Func<ModConfig> getConfig, Action reset, Action saveAndApply)
+    public GenericModConfigMenuIntegrationForTractor(IModRegistry modRegistry)
     {
         this.ModRegistry = modRegistry;
-        this.ConfigMenu = new GenericModConfigMenuIntegration<ModConfig>(modRegistry, monitor, manifest, getConfig, reset, saveAndApply);
     }
 
-    /// <summary>Register the config menu if available.</summary>
-    public void Register()
+    /// <inheritdoc />
+    public void Register(GenericModConfigMenuIntegration<ModConfig> menu, IMonitor monitor)
     {
         var defaultConfig = new ModConfig();
 
-        // get config menu
-        var menu = this.ConfigMenu;
-        if (!menu.IsLoaded)
-            return;
-
-        // register
         menu.Register();
 
         if (this.ModRegistry.IsLoaded("NermNermNerm.QuestableTractor"))